submission intermediate 2 - upload story

+++ app/src/main/java/com/example/storyapp/data/response/StoryResponse.kt
@@ -32,11 +32,11 @@ data class ListStoryItem(
 	val description: String,
 
 	@field:SerializedName("lon")
-	val lon: Double,
+	val lon: Float,
 
 	@field:SerializedName("id")
 	val id: String,
 
 	@field:SerializedName("lat")
-	val lat: Double
+	val lat: Float
 ): Parcelable
